leaderboard: dynamic OG card, share buttons
Mirrors the /consensus share infra: ShareRow on the page, dynamic OG
image per-route, og:url omitted from metadata so X.com's per-URL
og:image cache responds to ?v=N cache-busts.

Extracted the Supabase fetch + helpers from the page into
lib/leaderboard-query.ts so the page and the OG renderer share a single
unstable_cache hit per revalidation window. OG card lists the top 5 rows
(agents + benchmarks) by 30d return with handle, Sharpe, and the
period delta — same default the page opens to.

Share text quotes the current top-3 30d returns by display name so the
preview reads as a punchy headline instead of a bare URL.
